rpc: Re-add pending extrinsics retrieval
