+++ registry/error.go
@@ -13,6 +13,8 @@ const (
 	ErrEventsTypeNotFound                    = libErr.Error("events type not found")
 	ErrEventsTypeNotVariant                  = libErr.Error("events type not a variant")
 	ErrEventFieldsRetrieval                  = libErr.Error("event fields retrieval")
+	ErrExtrinsicFieldRetrieval               = libErr.Error("extrinsic field retrieval")
+	ErrInvalidExtrinsicParams                = libErr.Error("invalid extrinsic params")
 	ErrFieldDecoderForRecursiveFieldNotFound = libErr.Error("field decoder for recursive field not found")
 	ErrRecursiveFieldResolving               = libErr.Error("recursive field resolving")
 	ErrFieldTypeNotFound                     = libErr.Error("field type not found")
@@ -49,6 +51,10 @@ const (
 	ErrRecursiveFieldDecoderNotFound         = libErr.Error("recursive field decoder not found")
 	ErrBitVecDecoding                        = libErr.Error("bit vec decoding")
 	ErrNilTypeDecoder                        = libErr.Error("nil type decoder")
+	ErrNilExtrinsicDecoder                   = libErr.Error("nil type decoder")
+	ErrExtrinsicFieldNotFound                = libErr.Error("extrinsic field not found")
+	ErrExtrinsicCompactLengthDecoding        = libErr.Error("extrinsic compact length decoding")
+	ErrExtrinsicVersionDecoding              = libErr.Error("extrinsic version decoding")
 	ErrNilField                              = libErr.Error("nil field")
 	ErrNilFieldDecoder                       = libErr.Error("nil field decoder")
 	ErrFieldEncode                           = libErr.Error("field encoding")
