+++ registry/factory.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/centrifuge/go-substrate-rpc-client/v4/scale"
 	"github.com/centrifuge/go-substrate-rpc-client/v4/types"
 )
 
@@ -16,6 +15,7 @@ type Factory interface {
 	CreateCallRegistry(meta *types.Metadata) (CallRegistry, error)
 	CreateErrorRegistry(meta *types.Metadata) (ErrorRegistry, error)
 	CreateEventRegistry(meta *types.Metadata) (EventRegistry, error)
+	CreateExtrinsicDecoder(meta *types.Metadata) (*ExtrinsicDecoder, error)
 }
 
 // CallRegistry maps a call name to its TypeDecoder.
@@ -207,6 +207,44 @@ func (f *factory) CreateEventRegistry(meta *types.Metadata) (EventRegistry, erro
 	return eventRegistry, nil
 }
 
+const (
+	// ExpectedExtrinsicParams is the count of params that we expect for the Extrinsic metadata type.
+	//
+	// The parameters are expected to be in the following order:
+	// 1. Address
+	// 2. Call
+	// 3. Signature
+	// 4. Extra
+	ExpectedExtrinsicParams = 4
+)
+
+// CreateExtrinsicDecoder creates an ExtrinsicDecoder based on the Extrinsic information provided in the metadata.
+func (f *factory) CreateExtrinsicDecoder(meta *types.Metadata) (*ExtrinsicDecoder, error) {
+	f.resetStorages()
+
+	extrinsicLookupID := meta.AsMetadataV14.Extrinsic.Type
+
+	extrinsicType := meta.AsMetadataV14.EfficientLookup[extrinsicLookupID.Int64()]
+
+	if len(extrinsicType.Params) != ExpectedExtrinsicParams {
+		return nil, ErrInvalidExtrinsicParams
+	}
+
+	extrinsicFields, err := f.getTypeParams(meta, extrinsicType.Params)
+
+	if err != nil {
+		return nil, ErrExtrinsicFieldRetrieval
+	}
+
+	if err := f.resolveRecursiveDecoders(); err != nil {
+		return nil, ErrRecursiveDecodersResolving.Wrap(err)
+	}
+
+	return &ExtrinsicDecoder{
+		Fields: extrinsicFields,
+	}, nil
+}
+
 // resolveRecursiveDecoders resolves all recursive decoders with their according FieldDecoder.
 // nolint:lll
 func (f *factory) resolveRecursiveDecoders() error {
@@ -238,6 +276,48 @@ func (f *factory) resolveRecursiveDecoders() error {
 	}
 
 	return nil
+}
+
+// getTypeFields returns a list of fields and their respective decoders from the provided parameters.
+func (f *factory) getTypeParams(meta *types.Metadata, params []types.Si1TypeParameter) ([]*Field, error) {
+	var typeFields []*Field
+
+	for _, param := range params {
+		paramType, ok := meta.AsMetadataV14.EfficientLookup[param.Type.Int64()]
+
+		if !ok {
+			return nil, ErrFieldTypeNotFound.WithMsg(string(param.Name))
+		}
+
+		paramName := string(param.Name)
+
+		if storedFieldDecoder, ok := f.getStoredFieldDecoder(param.Type.Int64()); ok {
+			typeFields = append(typeFields, &Field{
+				Name:         paramName,
+				FieldDecoder: storedFieldDecoder,
+				LookupIndex:  param.Type.Int64(),
+			})
+			continue
+		}
+
+		paramTypeDef := paramType.Def
+
+		fieldDecoder, err := f.getFieldDecoder(meta, paramName, paramTypeDef)
+
+		if err != nil {
+			return nil, ErrFieldDecoderRetrieval.WithMsg(paramName).Wrap(err)
+		}
+
+		f.fieldStorage[param.Type.Int64()] = fieldDecoder
+
+		typeFields = append(typeFields, &Field{
+			Name:         paramName,
+			FieldDecoder: fieldDecoder,
+			LookupIndex:  param.Type.Int64(),
+		})
+	}
+
+	return typeFields, nil
 }
 
 // getTypeFields parses and returns all Field(s) for a type.
@@ -659,337 +739,4 @@ func getFieldName(field types.Si1Field) string {
 	default:
 		return fmt.Sprintf(lookupIndexFormat, field.Type.Int64())
 	}
-}
-
-// FieldDecoder is the interface implemented by all the different types that are available.
-type FieldDecoder interface {
-	Decode(decoder *scale.Decoder) (any, error)
-}
-
-// NoopDecoder is a FieldDecoder that does not decode anything. It comes in handy for nil tuples or variants
-// with no inner types.
-type NoopDecoder struct{}
-
-func (n *NoopDecoder) Decode(_ *scale.Decoder) (any, error) {
-	return nil, nil
-}
-
-// VariantDecoder holds a FieldDecoder for each variant/enum.
-type VariantDecoder struct {
-	FieldDecoderMap map[byte]FieldDecoder
-}
-
-func (v *VariantDecoder) Decode(decoder *scale.Decoder) (any, error) {
-	variantByte, err := decoder.ReadOneByte()
-
-	if err != nil {
-		return nil, ErrVariantByteDecoding.Wrap(err)
-	}
-
-	variantDecoder, ok := v.FieldDecoderMap[variantByte]
-
-	if !ok {
-		return nil, ErrVariantFieldDecoderNotFound.WithMsg("variant '%d'", variantByte)
-	}
-
-	if _, ok := variantDecoder.(*NoopDecoder); ok {
-		return variantByte, nil
-	}
-
-	return variantDecoder.Decode(decoder)
-}
-
-// ArrayDecoder holds information about the length of the array and the FieldDecoder used for its items.
-type ArrayDecoder struct {
-	Length      uint
-	ItemDecoder FieldDecoder
-}
-
-func (a *ArrayDecoder) Decode(decoder *scale.Decoder) (any, error) {
-	if a.ItemDecoder == nil {
-		return nil, ErrArrayItemDecoderNotFound
-	}
-
-	slice := make([]any, 0, a.Length)
-
-	for i := uint(0); i < a.Length; i++ {
-		item, err := a.ItemDecoder.Decode(decoder)
-
-		if err != nil {
-			return nil, ErrArrayItemDecoding.Wrap(err)
-		}
-
-		slice = append(slice, item)
-	}
-
-	return slice, nil
-}
-
-// SliceDecoder holds a FieldDecoder for the items of a vector/slice.
-type SliceDecoder struct {
-	ItemDecoder FieldDecoder
-}
-
-func (s *SliceDecoder) Decode(decoder *scale.Decoder) (any, error) {
-	if s.ItemDecoder == nil {
-		return nil, ErrSliceItemDecoderNotFound
-	}
-
-	sliceLen, err := decoder.DecodeUintCompact()
-
-	if err != nil {
-		return nil, ErrSliceLengthDecoding.Wrap(err)
-	}
-
-	slice := make([]any, 0, sliceLen.Uint64())
-
-	for i := uint64(0); i < sliceLen.Uint64(); i++ {
-		item, err := s.ItemDecoder.Decode(decoder)
-
-		if err != nil {
-			return nil, ErrSliceItemDecoding.Wrap(err)
-		}
-
-		slice = append(slice, item)
-	}
-
-	return slice, nil
-}
-
-// CompositeDecoder holds all the information required to decoder a struct/composite.
-type CompositeDecoder struct {
-	FieldName string
-	Fields    []*Field
-}
-
-func (e *CompositeDecoder) Decode(decoder *scale.Decoder) (any, error) {
-	var decodedFields DecodedFields
-
-	for _, field := range e.Fields {
-		value, err := field.FieldDecoder.Decode(decoder)
-
-		if err != nil {
-			return nil, ErrCompositeFieldDecoding.Wrap(err)
-		}
-
-		decodedFields = append(decodedFields, &DecodedField{
-			Name:        field.Name,
-			Value:       value,
-			LookupIndex: field.LookupIndex,
-		})
-	}
-
-	return decodedFields, nil
-}
-
-// ValueDecoder decodes a primitive type.
-type ValueDecoder[T any] struct{}
-
-func (v *ValueDecoder[T]) Decode(decoder *scale.Decoder) (any, error) {
-	var t T
-
-	if err := decoder.Decode(&t); err != nil {
-		return nil, ErrValueDecoding.Wrap(err)
-	}
-
-	return t, nil
-}
-
-// RecursiveDecoder is a wrapper for a FieldDecoder that is recursive.
-type RecursiveDecoder struct {
-	FieldDecoder FieldDecoder
-}
-
-func (r *RecursiveDecoder) Decode(decoder *scale.Decoder) (any, error) {
-	if r.FieldDecoder == nil {
-		return nil, ErrRecursiveFieldDecoderNotFound
-	}
-
-	return r.FieldDecoder.Decode(decoder)
-}
-
-// BitSequenceDecoder holds decoding information for a bit sequence.
-type BitSequenceDecoder struct {
-	FieldName string
-	BitOrder  types.BitOrder
-}
-
-func (b *BitSequenceDecoder) Decode(decoder *scale.Decoder) (any, error) {
-	bitVec := types.NewBitVec(b.BitOrder)
-
-	if err := bitVec.Decode(*decoder); err != nil {
-		return nil, ErrBitVecDecoding.Wrap(err)
-	}
-
-	return map[string]string{
-		b.FieldName: bitVec.String(),
-	}, nil
-}
-
-// TypeDecoder holds all information required to decode a particular type.
-type TypeDecoder struct {
-	Name   string
-	Fields []*Field
-}
-
-func (t *TypeDecoder) Decode(decoder *scale.Decoder) (DecodedFields, error) {
-	if t == nil {
-		return nil, ErrNilTypeDecoder
-	}
-
-	var decodedFields DecodedFields
-
-	for _, field := range t.Fields {
-		decodedField, err := field.Decode(decoder)
-
-		if err != nil {
-			return nil, ErrTypeFieldDecoding.Wrap(err)
-		}
-
-		decodedFields = append(decodedFields, decodedField)
-	}
-
-	return decodedFields, nil
-}
-
-// Field represents one field of a TypeDecoder.
-type Field struct {
-	Name         string
-	FieldDecoder FieldDecoder
-	LookupIndex  int64
-}
-
-func (f *Field) Decode(decoder *scale.Decoder) (*DecodedField, error) {
-	if f == nil {
-		return nil, ErrNilField
-	}
-
-	if f.FieldDecoder == nil {
-		return nil, ErrNilFieldDecoder
-	}
-
-	value, err := f.FieldDecoder.Decode(decoder)
-
-	if err != nil {
-		return nil, err
-	}
-
-	return &DecodedField{
-		Name:        f.Name,
-		Value:       value,
-		LookupIndex: f.LookupIndex,
-	}, nil
-}
-
-// DecodedField holds the name, value and lookup index of a field that was decoded.
-type DecodedField struct {
-	Name        string
-	Value       any
-	LookupIndex int64
-}
-
-func (d DecodedField) Encode(encoder scale.Encoder) error {
-	if d.Value == nil {
-		return nil
-	}
-
-	return encoder.Encode(d.Value)
-}
-
-type DecodedFields []*DecodedField
-
-type DecodedFieldPredicateFn func(fieldIndex int, field *DecodedField) bool
-type DecodedValueProcessingFn[T any] func(value any) (T, error)
-
-// ProcessDecodedFieldValue applies the processing func to the value of the field
-// that matches the provided predicate func.
-func ProcessDecodedFieldValue[T any](
-	decodedFields DecodedFields,
-	fieldPredicateFn DecodedFieldPredicateFn,
-	valueProcessingFn DecodedValueProcessingFn[T],
-) (T, error) {
-	var t T
-
-	for decodedFieldIndex, decodedField := range decodedFields {
-		if !fieldPredicateFn(decodedFieldIndex, decodedField) {
-			continue
-		}
-
-		res, err := valueProcessingFn(decodedField.Value)
-
-		if err != nil {
-			return t, ErrDecodedFieldValueProcessingError.Wrap(err)
-		}
-
-		return res, nil
-	}
-
-	return t, ErrDecodedFieldNotFound
-}
-
-// GetDecodedFieldAsType returns the value of the field that matches the provided predicate func
-// as the provided generic argument.
-func GetDecodedFieldAsType[T any](
-	decodedFields DecodedFields,
-	fieldPredicateFn DecodedFieldPredicateFn,
-) (T, error) {
-	return ProcessDecodedFieldValue(
-		decodedFields,
-		fieldPredicateFn,
-		func(value any) (T, error) {
-			if res, ok := value.(T); ok {
-				return res, nil
-			}
-
-			var t T
-
-			err := fmt.Errorf("expected %T, got %T", t, value)
-
-			return t, ErrDecodedFieldValueTypeMismatch.Wrap(err)
-		},
-	)
-}
-
-// GetDecodedFieldAsSliceOfType returns the value of the field that matches the provided predicate func
-// as a slice of the provided generic argument.
-func GetDecodedFieldAsSliceOfType[T any](
-	decodedFields DecodedFields,
-	fieldPredicateFn DecodedFieldPredicateFn,
-) ([]T, error) {
-	return ProcessDecodedFieldValue(
-		decodedFields,
-		fieldPredicateFn,
-		func(value any) ([]T, error) {
-			v, ok := value.([]any)
-
-			if !ok {
-				return nil, ErrDecodedFieldValueNotAGenericSlice
-			}
-
-			res, err := convertSliceToType[T](v)
-
-			if err != nil {
-				return nil, ErrDecodedFieldValueTypeMismatch.Wrap(err)
-			}
-
-			return res, nil
-		},
-	)
-}
-
-func convertSliceToType[T any](slice []any) ([]T, error) {
-	res := make([]T, 0)
-
-	for _, item := range slice {
-		if v, ok := item.(T); ok {
-			res = append(res, v)
-			continue
-		}
-
-		var t T
-
-		return nil, fmt.Errorf("expected %T, got %T", t, item)
-	}
-
-	return res, nil
 }
